Extract book assets
